refactor: 폴더 구조 정리 및 index.tsx 간소화
