+++ frontend/src/pages/Todo/index.tsx
@@ -1,695 +1,64 @@
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-  type Dispatch,
-  type SetStateAction,
-} from "react";
-import { isAxiosError } from "axios";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Header } from "@/components/Header";
-import { QUERY_KEYS } from "@/apis/constants/queryKeys";
-import { getStoredMember } from "@/apis/services/auth";
-import { todoApi } from "@/apis/services/todo";
-import type {
-  CreateTodoBody,
-  TodoApiItem,
-  TodoCategoryApiItem,
-  UpdateTodoBody,
-} from "@/apis/types/todo";
-import EditCard from "@/pages/Todo/Edit/EditCard";
-import TodoList from "@/pages/Todo/TodoList/TodoList";
-import TodoToast, {
-  type TodoToastPayload,
-} from "@/pages/Todo/TodoToast";
-import {
-  DEFAULT_CATEGORIES,
-  type Category,
-  type TodoItem,
-} from "@/pages/Todo/types";
-
-function mapTodoCategoryApiToCategory(row: TodoCategoryApiItem): Category {
-  return { id: String(row.id), label: row.name };
-}
-
-function mapTodoApiToItem(row: TodoApiItem): TodoItem {
-  return {
-    id: String(row.id),
-    title: row.content,
-    completed: row.isDone,
-    dueDate: row.dueDate,
-    categoryId: row.categoryId != null ? String(row.categoryId) : "",
-    categoryName: row.categoryName,
-  };
-}
-
-function buildCreateTodoBody(item: {
-  title: string;
-  dueDate: string;
-  categoryId: string;
-}): CreateTodoBody {
-  const body: CreateTodoBody = {
-    content: item.title.trim(),
-    dueDate: item.dueDate.trim(),
-  };
-  const raw = item.categoryId?.trim();
-  if (raw && !raw.startsWith("opt-")) {
-    const n = Number(raw);
-    if (Number.isFinite(n) && n > 0) {
-      body.categoryId = n;
-    }
-  }
-  return body;
-}
-
-function buildUpdateTodoBody(
-  prev: TodoItem,
-  item: { title: string; dueDate: string; categoryId: string },
-): UpdateTodoBody {
-  const body: UpdateTodoBody = {
-    content: item.title.trim(),
-    dueDate: item.dueDate.trim(),
-    isDone: prev.completed,
-  };
-  const raw = item.categoryId?.trim();
-  if (!raw) {
-    body.removeCategory = true;
-    return body;
-  }
-  if (raw.startsWith("opt-")) {
-    body.removeCategory = true;
-    return body;
-  }
-  const n = Number(raw);
-  if (Number.isFinite(n) && n > 0) {
-    body.removeCategory = false;
-    body.categoryId = n;
-  } else {
-    body.removeCategory = true;
-  }
-  return body;
-}
-
-function buildOptimisticTodoItem(
-  item: { title: string; dueDate: string; categoryId: string },
-  tempId: string,
-  categories: Category[],
-): TodoItem {
-  const raw = item.categoryId?.trim() ?? "";
-  let categoryName: string | null | undefined;
-  if (raw && !raw.startsWith("opt-")) {
-    const cat = categories.find((c) => c.id === raw);
-    categoryName = cat?.label ?? undefined;
-  }
-  return {
-    id: tempId,
-    title: item.title.trim(),
-    dueDate: item.dueDate.trim(),
-    categoryId: raw,
-    categoryName,
-    completed: false,
-  };
-}
-
-function applyTodoUpdateOptimistic(
-  prev: TodoItem,
-  body: UpdateTodoBody,
-  categories: Category[],
-): TodoItem {
-  let categoryId = prev.categoryId;
-  let categoryName = prev.categoryName ?? null;
-
-  if (body.removeCategory) {
-    categoryId = "";
-    categoryName = null;
-  } else if (body.categoryId !== undefined) {
-    categoryId = String(body.categoryId);
-    const cat = categories.find((c) => c.id === categoryId);
-    categoryName = cat?.label ?? null;
-  }
-
-  return {
-    ...prev,
-    title: body.content !== undefined ? body.content : prev.title,
-    completed: body.isDone !== undefined ? body.isDone : prev.completed,
-    dueDate: body.dueDate !== undefined ? body.dueDate : prev.dueDate,
-    categoryId,
-    categoryName,
-  };
-}
+import EditCard from "@/pages/Todo/components/Edit/EditCard";
+import TodoList from "@/pages/Todo/components/TodoList";
+import TodoToast from "@/pages/Todo/components/TodoToast";
+import { useTodoBoard } from "@/pages/Todo/components/TodoBoard/useTodoBoard";
 
 const TodoPage = () => {
-  const queryClient = useQueryClient();
-  const userId = getStoredMember()?.id ?? null;
-
-  const {
-    data: serverTodos = [],
-    isLoading: isTodosLoading,
-    isError: isTodosError,
-  } = useQuery({
-    queryKey:
-      userId != null
-        ? QUERY_KEYS.todos.byUser(userId)
-        : (["todos", "guest"] as const),
-    queryFn: async () => {
-      const rows = await todoApi.getListByUserId(userId!);
-      return rows.map(mapTodoApiToItem);
-    },
-    enabled: userId != null,
-  });
-
-  const [guestTodos, setGuestTodos] = useState<TodoItem[]>([]);
-  const todos = userId != null ? serverTodos : guestTodos;
-
-  const {
-    data: serverCategories = [],
-    isLoading: isCategoriesLoading,
-    isError: isCategoriesError,
-  } = useQuery({
-    queryKey:
-      userId != null
-        ? QUERY_KEYS.todoCategories.byUser(userId)
-        : (["todoCategories", "guest"] as const),
-    queryFn: async () => {
-      const rows = await todoApi.getCategoriesByUserId(userId!);
-      return rows.map(mapTodoCategoryApiToCategory);
-    },
-    enabled: userId != null,
-  });
-
-  const [guestCategories, setGuestCategories] = useState<Category[]>(
-    DEFAULT_CATEGORIES,
-  );
-
-  const categories = userId != null ? serverCategories : guestCategories;
-
-  const setCategories: Dispatch<SetStateAction<Category[]>> = (action) => {
-    if (userId == null) {
-      setGuestCategories(action);
-      return;
-    }
-    queryClient.setQueryData<Category[]>(
-      QUERY_KEYS.todoCategories.byUser(userId),
-      (prev) => {
-        const p = prev ?? [];
-        return typeof action === "function" ? action(p) : action;
-      },
-    );
-  };
-
-  const patchTodos = useCallback(
-    (updater: (prev: TodoItem[]) => TodoItem[]) => {
-      if (userId == null) {
-        setGuestTodos((prev) => updater(prev));
-        return;
-      }
-      queryClient.setQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-        (prev) => updater(prev ?? []),
-      );
-    },
-    [userId, queryClient],
-  );
-
-  const [categoryRemap, setCategoryRemap] = useState<{
-    tempId: string;
-    realId: string;
-  } | null>(null);
-
-  const [categoryCreateFailedTempId, setCategoryCreateFailedTempId] =
-    useState<string | null>(null);
-
-  const onCategoryRemapConsumed = useCallback(() => {
-    setCategoryRemap(null);
-  }, []);
-
-  const onCategoryCreateFailedConsumed = useCallback(() => {
-    setCategoryCreateFailedTempId(null);
-  }, []);
-
-  const [todoToast, setTodoToast] = useState<TodoToastPayload | null>(null);
-
-  const clearTodoToast = useCallback(() => setTodoToast(null), []);
-
-  const notifyTodo = useCallback(
-    (text: string, variant: "success" | "error" = "success") => {
-      setTodoToast((prev) => ({
-        key: (prev?.key ?? 0) + 1,
-        text,
-        variant,
-      }));
-    },
-    [],
-  );
-
-  const createCategoryMutation = useMutation({
-    mutationFn: ({
-      name,
-      tempId: _tempId,
-    }: {
-      name: string;
-      tempId: string;
-    }) => todoApi.createCategory(userId!, { name }),
-    onMutate: async ({ name, tempId }) => {
-      if (userId == null) return {};
-      await queryClient.cancelQueries({
-        queryKey: QUERY_KEYS.todoCategories.byUser(userId),
-      });
-      const previous = queryClient.getQueryData<Category[]>(
-        QUERY_KEYS.todoCategories.byUser(userId),
-      );
-      queryClient.setQueryData<Category[]>(
-        QUERY_KEYS.todoCategories.byUser(userId),
-        (prev) => [...(prev ?? []), { id: tempId, label: name }],
-      );
-      return { previous };
-    },
-    onError: (err, { tempId }, context) => {
-      if (userId == null) return;
-      if (context?.previous !== undefined) {
-        queryClient.setQueryData(
-          QUERY_KEYS.todoCategories.byUser(userId),
-          context.previous,
-        );
-      } else {
-        queryClient.setQueryData<Category[]>(
-          QUERY_KEYS.todoCategories.byUser(userId),
-          (prev) => (prev ?? []).filter((c) => c.id !== tempId),
-        );
-      }
-      setCategoryCreateFailedTempId(tempId);
-      if (isAxiosError(err)) {
-        const status = err.response?.status;
-        if (status === 409) {
-          notifyTodo("같은 이름의 카테고리가 이미 있어요.", "error");
-        } else {
-          notifyTodo("카테고리를 추가하지 못했어요.", "error");
-        }
-      } else {
-        notifyTodo("카테고리를 추가하지 못했어요.", "error");
-      }
-    },
-    onSuccess: (data, { tempId }) => {
-      if (userId == null) return;
-      queryClient.setQueryData<Category[]>(
-        QUERY_KEYS.todoCategories.byUser(userId),
-        (prev) =>
-          (prev ?? []).map((c) =>
-            c.id === tempId ? mapTodoCategoryApiToCategory(data) : c,
-          ),
-      );
-      setCategoryRemap({ tempId, realId: String(data.id) });
-    },
-  });
-
-  const handleCreateCategory = (name: string): { tempId: string } => {
-    if (userId == null) return { tempId: "" };
-    const trimmed = name.trim().slice(0, 50);
-    if (!trimmed) return { tempId: "" };
-    const tempId = `opt-${crypto.randomUUID()}`;
-    createCategoryMutation.mutate({ name: trimmed, tempId });
-    return { tempId };
-  };
-
-  const createTodoMutation = useMutation({
-    mutationFn: (vars: {
-      item: {
-        title: string;
-        dueDate: string;
-        categoryId: string;
-      };
-      tempId: string;
-    }) => todoApi.createTodo(userId!, buildCreateTodoBody(vars.item)),
-    onMutate: async ({ item, tempId }) => {
-      if (userId == null) return {};
-      await queryClient.cancelQueries({
-        queryKey: QUERY_KEYS.todos.byUser(userId),
-      });
-      const previous = queryClient.getQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-      );
-      const cats =
-        queryClient.getQueryData<Category[]>(
-          QUERY_KEYS.todoCategories.byUser(userId),
-        ) ?? [];
-      queryClient.setQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-        (prev) => [
-          ...(prev ?? []),
-          buildOptimisticTodoItem(item, tempId, cats),
-        ],
-      );
-      return { previous };
-    },
-    onSuccess: (data, { tempId }) => {
-      if (userId == null) return;
-      queryClient.setQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-        (prev) => {
-          const list = prev ?? [];
-          const mapped = mapTodoApiToItem(data);
-          if (!list.some((t) => t.id === tempId)) {
-            return [...list, mapped];
-          }
-          return list.map((t) => (t.id === tempId ? mapped : t));
-        },
-      );
-      notifyTodo("생성됐어요");
-    },
-    onError: (err, _v, context) => {
-      if (userId == null) return;
-      if (context?.previous !== undefined) {
-        queryClient.setQueryData(
-          QUERY_KEYS.todos.byUser(userId),
-          context.previous,
-        );
-      }
-      if (isAxiosError(err)) {
-        const status = err.response?.status;
-        if (status === 400) {
-          notifyTodo("입력값을 확인해 주세요.", "error");
-        } else if (status === 403) {
-          notifyTodo("권한이 없어요.", "error");
-        } else if (status === 404) {
-          notifyTodo("사용자 또는 카테고리를 찾을 수 없어요.", "error");
-        } else {
-          notifyTodo("할 일을 추가하지 못했어요.", "error");
-        }
-      } else {
-        notifyTodo("할 일을 추가하지 못했어요.", "error");
-      }
-    },
-  });
-
-  const updateTodoMutation = useMutation({
-    mutationFn: ({
-      todoId,
-      body,
-    }: {
-      todoId: number;
-      todoIdStr: string;
-      body: UpdateTodoBody;
-    }) => todoApi.updateTodo(todoId, body),
-    onMutate: async ({ body, todoIdStr }) => {
-      if (userId == null) return {};
-      await queryClient.cancelQueries({
-        queryKey: QUERY_KEYS.todos.byUser(userId),
-      });
-      const previous = queryClient.getQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-      );
-      const cats =
-        queryClient.getQueryData<Category[]>(
-          QUERY_KEYS.todoCategories.byUser(userId),
-        ) ?? [];
-      const prevTodo = previous?.find((t) => t.id === todoIdStr);
-      if (!prevTodo) return { previous };
-      const next = applyTodoUpdateOptimistic(prevTodo, body, cats);
-      queryClient.setQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-        (prev) =>
-          (prev ?? []).map((t) => (t.id === todoIdStr ? next : t)),
-      );
-      setEditingId(null);
-      return { previous };
-    },
-    onSuccess: (data) => {
-      if (userId == null) return;
-      queryClient.setQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-        (prev) =>
-          (prev ?? []).map((t) =>
-            t.id === String(data.id) ? mapTodoApiToItem(data) : t,
-          ),
-      );
-      setEditingId(null);
-      notifyTodo("수정됐어요");
-    },
-    onError: (err, _v, context) => {
-      if (userId == null) return;
-      if (context?.previous !== undefined) {
-        queryClient.setQueryData(
-          QUERY_KEYS.todos.byUser(userId),
-          context.previous,
-        );
-      }
-      if (isAxiosError(err)) {
-        const status = err.response?.status;
-        if (status === 400) {
-          notifyTodo("내용은 비울 수 없어요.", "error");
-        } else if (status === 403) {
-          notifyTodo("권한이 없어요.", "error");
-        } else if (status === 404) {
-          notifyTodo("할 일 또는 카테고리를 찾을 수 없어요.", "error");
-        } else {
-          notifyTodo("수정에 실패했어요.", "error");
-        }
-      } else {
-        notifyTodo("수정에 실패했어요.", "error");
-      }
-    },
-  });
-
-  const deleteTodoMutation = useMutation({
-    mutationFn: async ({
-      todoId,
-    }: {
-      todoId: number;
-      todoIdStr: string;
-    }) => {
-      await todoApi.deleteTodo(todoId);
-    },
-    onMutate: async ({ todoIdStr }) => {
-      if (userId == null) return {};
-      await queryClient.cancelQueries({
-        queryKey: QUERY_KEYS.todos.byUser(userId),
-      });
-      const previous = queryClient.getQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-      );
-      queryClient.setQueryData<TodoItem[]>(
-        QUERY_KEYS.todos.byUser(userId),
-        (prev) => (prev ?? []).filter((t) => t.id !== todoIdStr),
-      );
-      setEditingId((cur) => (cur === todoIdStr ? null : cur));
-      return { previous };
-    },
-    onError: (err, _vars, context) => {
-      if (userId == null) return;
-      if (context?.previous !== undefined) {
-        queryClient.setQueryData(
-          QUERY_KEYS.todos.byUser(userId),
-          context.previous,
-        );
-      }
-      if (isAxiosError(err)) {
-        const status = err.response?.status;
-        if (status === 403 || status === 404) {
-          notifyTodo("할 일을 삭제할 수 없어요.", "error");
-        } else {
-          notifyTodo("삭제에 실패했어요.", "error");
-        }
-      } else {
-        notifyTodo("삭제에 실패했어요.", "error");
-      }
-    },
-    onSuccess: () => {
-      notifyTodo("삭제됐어요");
-    },
-  });
-
-  const deleteCategoryMutation = useMutation({
-    mutationFn: async ({
-      categoryId,
-    }: {
-      categoryId: number;
-      categoryIdStr: string;
-    }) => {
-      await todoApi.deleteCategory(userId!, categoryId);
-    },
-    onMutate: async ({ categoryIdStr }) => {
-      if (userId == null) return {};
-      await queryClient.cancelQueries({
-        queryKey: QUERY_KEYS.todoCategories.byUser(userId),
-      });
-      const previous = queryClient.getQueryData<Category[]>(
-        QUERY_KEYS.todoCategories.byUser(userId),
-      );
-      queryClient.setQueryData<Category[]>(
-        QUERY_KEYS.todoCategories.byUser(userId),
-        (prev) => (prev ?? []).filter((c) => c.id !== categoryIdStr),
-      );
-      return { previous };
-    },
-    onError: (err, _vars, context) => {
-      if (userId == null) return;
-      if (context?.previous !== undefined) {
-        queryClient.setQueryData(
-          QUERY_KEYS.todoCategories.byUser(userId),
-          context.previous,
-        );
-      }
-      if (isAxiosError(err)) {
-        const status = err.response?.status;
-        if (status === 403 || status === 404) {
-          notifyTodo("카테고리를 삭제할 수 없어요.", "error");
-        } else {
-          notifyTodo("카테고리 삭제에 실패했어요.", "error");
-        }
-      } else {
-        notifyTodo("카테고리 삭제에 실패했어요.", "error");
-      }
-    },
-  });
-
-  const handleRemoveCategory = useCallback(
-    (id: string) => {
-      if (userId == null) {
-        setGuestCategories((prev) => prev.filter((c) => c.id !== id));
-        return;
-      }
-      if (id.startsWith("opt-")) {
-        queryClient.setQueryData<Category[]>(
-          QUERY_KEYS.todoCategories.byUser(userId),
-          (prev) => (prev ?? []).filter((c) => c.id !== id),
-        );
-        return;
-      }
-      const numId = Number(id);
-      if (!Number.isFinite(numId) || numId <= 0) return;
-      deleteCategoryMutation.mutate({ categoryId: numId, categoryIdStr: id });
-    },
-    [userId, queryClient, deleteCategoryMutation],
-  );
-
-  const [editingId, setEditingId] = useState<string | null>(null);
-
-  const editingTodo = useMemo(
-    () =>
-      editingId ? (todos.find((t) => t.id === editingId) ?? null) : null,
-    [editingId, todos],
-  );
-
-  useEffect(() => {
-    if (editingId && !todos.some((t) => t.id === editingId)) {
-      setEditingId(null);
-    }
-  }, [editingId, todos]);
-
-  const handleAddTodo = (item: {
-    title: string;
-    dueDate: string;
-    categoryId: string;
-  }) => {
-    if (userId == null) {
-      patchTodos((prev) => [
-        ...prev,
-        { id: crypto.randomUUID(), ...item, completed: false },
-      ]);
-      notifyTodo("생성됐어요");
-      return;
-    }
-    const tempId = `opt-todo-${crypto.randomUUID()}`;
-    createTodoMutation.mutate({ item, tempId });
-  };
-
-  const handleUpdateTodo = (
-    id: string,
-    item: { title: string; dueDate: string; categoryId: string },
-  ) => {
-    if (userId == null) {
-      patchTodos((prev) =>
-        prev.map((t) => (t.id === id ? { ...t, ...item } : t)),
-      );
-      setEditingId(null);
-      notifyTodo("수정됐어요");
-      return;
-    }
-    const prevTodo = todos.find((t) => t.id === id);
-    if (!prevTodo) return;
-    const todoIdNum = Number(id);
-    if (!Number.isFinite(todoIdNum) || todoIdNum <= 0) {
-      patchTodos((prev) =>
-        prev.map((t) => (t.id === id ? { ...t, ...item } : t)),
-      );
-      setEditingId(null);
-      notifyTodo("수정됐어요");
-      return;
-    }
-    updateTodoMutation.mutate({
-      todoId: todoIdNum,
-      todoIdStr: id,
-      body: buildUpdateTodoBody(prevTodo, item),
-    });
-  };
-
-  const handleDeleteTodo = (id: string) => {
-    if (userId == null) {
-      patchTodos((prev) => prev.filter((t) => t.id !== id));
-      setEditingId(null);
-      notifyTodo("삭제됐어요");
-      return;
-    }
-    const num = Number(id);
-    if (!Number.isFinite(num) || num <= 0) {
-      patchTodos((prev) => prev.filter((t) => t.id !== id));
-      setEditingId(null);
-      notifyTodo("삭제됐어요");
-      return;
-    }
-    deleteTodoMutation.mutate({ todoId: num, todoIdStr: id });
-  };
+  const b = useTodoBoard();
 
   return (
     <div className="relative flex h-screen w-full flex-col overflow-hidden bg-gray-darkest pt-[65px]">
-      <TodoToast toast={todoToast} onClear={clearTodoToast} />
       <Header variant="dark" alwaysVisible />
-      <div className="flex min-h-0 flex-1 flex-col gap-2 px-8 pb-8 pt-4">
-        {userId != null && isCategoriesError ? (
-          <p className="shrink-0 text-sm text-red-400/90">
-            카테고리 목록을 불러오지 못했습니다.
-          </p>
-        ) : null}
-        <div className="flex min-h-0 min-w-0 flex-1 flex-row gap-3">
-          <TodoList
-            todos={todos}
-            categories={categories}
-            isLoading={
-              Boolean(userId) && (isTodosLoading || isCategoriesLoading)
-            }
-            isError={isTodosError}
-            onEditTodo={(todo) => setEditingId(todo.id)}
-            onToggleComplete={(id) =>
-              patchTodos((prev) =>
-                prev.map((t) =>
-                  t.id === id ? { ...t, completed: !t.completed } : t,
-                ),
-              )
-            }
-            onMoveTodoToDate={(id, dueDate) =>
-              patchTodos((prev) =>
-                prev.map((t) => (t.id === id ? { ...t, dueDate } : t)),
-              )
-            }
-          />
+      <div className="flex min-h-0 min-w-0 flex-1 flex-col">
+        <TodoToast toast={b.todoToast} onClear={b.clearTodoToast} />
+        <div className="flex min-h-0 flex-1 flex-col gap-2 px-8 pb-8 pt-4">
+          {b.userId != null && b.isCategoriesError ? (
+            <p className="shrink-0 text-sm text-red-400/90">
+              카테고리 목록을 불러오지 못했습니다.
+            </p>
+          ) : null}
+          <div className="flex min-h-0 min-w-0 flex-1 flex-row gap-3">
+            <TodoList
+              todos={b.todos}
+              categories={b.categories}
+              isLoading={
+                Boolean(b.userId) &&
+                (b.isTodosLoading || b.isCategoriesLoading)
+              }
+              isError={b.isTodosError}
+              onEditTodo={(todo) => b.setEditingId(todo.id)}
+              onToggleComplete={(id) =>
+                b.patchTodos((prev) =>
+                  prev.map((t) =>
+                    t.id === id ? { ...t, completed: !t.completed } : t,
+                  ),
+                )
+              }
+              onMoveTodoToDate={(id, dueDate) =>
+                b.patchTodos((prev) =>
+                  prev.map((t) => (t.id === id ? { ...t, dueDate } : t)),
+                )
+              }
+            />
 
-          <EditCard
-            categories={categories}
-            setCategories={setCategories}
-            onRemoveCategory={handleRemoveCategory}
-            onCreateCategory={
-              userId != null ? handleCreateCategory : undefined
-            }
-            categoryRemap={categoryRemap}
-            onCategoryRemapConsumed={onCategoryRemapConsumed}
-            categoryCreateFailedTempId={categoryCreateFailedTempId}
-            onCategoryCreateFailedConsumed={onCategoryCreateFailedConsumed}
-            editingTodo={editingTodo}
-            onAddTodo={handleAddTodo}
-            onUpdateTodo={handleUpdateTodo}
-            onDeleteTodo={handleDeleteTodo}
-          />
+            <EditCard
+              categories={b.categories}
+              setCategories={b.setCategories}
+              onRemoveCategory={b.handleRemoveCategory}
+              onCreateCategory={
+                b.userId != null ? b.handleCreateCategory : undefined
+              }
+              categoryRemap={b.categoryRemap}
+              onCategoryRemapConsumed={b.onCategoryRemapConsumed}
+              categoryCreateFailedTempId={b.categoryCreateFailedTempId}
+              onCategoryCreateFailedConsumed={b.onCategoryCreateFailedConsumed}
+              editingTodo={b.editingTodo}
+              onAddTodo={b.handleAddTodo}
+              onUpdateTodo={b.handleUpdateTodo}
+              onDeleteTodo={b.handleDeleteTodo}
+            />
+          </div>
         </div>
       </div>
     </div>
